Change the starting point of the app
